requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
PyYAML>=6.0
PyQt6>=6.4.0

# Optional ML dependencies (uncomment if needed)
# tensorflow>=2.8.0
# torch>=1.10.0
# torchvision>=0.11.0
# torchaudio>=0.10.0

asttokens==2.4.1
backcall==0.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
executing==2.0.1
filelock==3.15.4
fonttools==4.53.0
fsspec==2024.6.1
idna==3.7
imageio==2.34.2
imgaug==0.4.0
importlib-metadata==8.0.0
importlib-resources==6.4.0
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.1
jinja2>=3.1.5
jupyter-client==8.6.2
jupyter-core==5.7.2
kiwisolver==1.4.5
lazy-loader==0.4
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mkl==2021.4.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
packaging==24.1
pandas==2.0.3
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
prompt-toolkit==3.0.47
psutil==6.0.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.4.1
pywin32==306
requests==2.32.3
scikit-image==0.21.0
scikit-learn>=1.0.2
scipy==1.10.1
seaborn==0.13.2
shapely==2.0.4
six==1.16.0
stack-data==0.6.3
sympy==1.12.1
tbb==2021.13.0
tifffile==2023.7.10
tornado>=6.4.2
tqdm==4.66.4
traitlets==5.14.3
treys>=0.1.8
typing-extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.49
ultralytics-thop==2.0.0
urllib3==2.2.2
wcwidth==0.2.13
zipp==3.19.2
